reorganize gguf utils
